test/ops_replayer: create separate ops_parser
