+++ src/test/objectstore/allocsim/CMakeLists.txt
@@ -1,6 +1,6 @@
 
 
-add_executable(replayer ops_replayer.cc)
+add_executable(replayer ops_replayer.cc ops_parser.cc)
 
 target_link_libraries(replayer
     PRIVATE
